api/requirements.txt
flask
flask-cors
nltk
textblob
twilio
python-dotenv
